Reimplement mapAsyncPartitioned() - again - this time as a StepSource.
It turns out we can implement this using any StepSource as the first 'buffer', and have the workers drain from that. This removes an unnecessary buffer, and distills more of the essence of this operator, which turns out to not be a standard shape (it is Sink -> StepSource). To accommodate this, the operator accepts an 'upstream' StepSource, and connects it to the balanced worker sink in the run() method.