Overhaul interfaces to support all 9 possible shapes as first-class citizens, with arbitrary chaining between them.
This comes after implementing mapAsyncPartitioned(), which I had initially hacked into a StepSource for lack of a natural SinkStepSource. This felt particularly uncanny given that the simpler version - mapAsync() - was a Sink, thus nonsensically appeared diametrically opposite. I initially thought it would unbearably increase complexity to add the combinatoric interfaces, along with all the chaining methods. But in the end I think I accomplished it in a way that keeps the mental model small, while minimizing footguns. The full set of Stages is now: Sink, Source, StepSink, StepSource, Silo (new). The 4 remaining interfaces are combinations of 'SinkSource', including the pre-existing Segue. For practicality, chaining is aware of these 4 interfaces, but otherwise they are merely containers for a sink() and source().