Add a new interface (Forker) to replace Consumer-of-Callable.
I really wish I could use Executor here, but it accepts Runnable, which cannot throw checked exceptions. Maybe it would be possible to redesign so that call sites do not throw checked exceptions - specifically, the Sink.complete(x) in Silo.run() would have to either not throw or be caught and handled somehow. But I'm not there yet.