Bifurcate Sink#complete(Throwable) into Sink#complete() (throws Exception) and Sink#completeExceptionally(Throwable) (does not throw).
I avoided this for so long because it obviously adds more surface area for implementors. But previously changing run() to accept an Executor (which accepts a Runnable, which can't throw Exception) added pressure to the issue. Now that I've actually done it, I believe this segregation actually makes implementation of Segues easier to reason about: complete() should prevent further offers; completeExceptionally() should do that AND throw on further polls. This also makes implementing a TimedSegue/Core easier, since completeExceptionally() can be and now is fully implemented internally in TimedSegue - Core implementors do not need to fuss with error propagation.