Switch nested operators to use a real STS, instead of asynchronously capturing + wrapping the Executor from run().
The downside here is that nested operators now have to take an 'innerExceptionHandler', else there would be no way to observe exceptions thrown from tasks in the inner scope. (This was not a problem previously, because the tasks were run in the captured executor. This also meant that exceptions might have been captured even after the inner 'scope' was shutdown, which is no longer the case.) The upside is that we get real thread structuring in named scopes, and there is no more deadlock-scary async executor capture.