Re-add 'throws Exception' to Sink#completeExceptionally, and change the default impl to do nothing (instead of throwing CompletionException).
Key insight is that most places that call completeExceptionally are themselves completeExceptionally, and are fine to throw exceptions from completing downstream sinks. The only places that differ are the 'root callers', that originate the exception that gets passed to completeExceptionally. This is mostly ClosedSilo#run(), which can suppress any exception from completeExceptionally onto the original exception, then wrap the original exception in CompletionException (an unchecked exception type, since it is called by an Executor and cannot throw a checked Exception).