Remove 'called' checks from plain Sinks, and remove state-handling from some complete/close methods.
The 'called' checks appear to merely soften the effect of improper/concurrent calls, at the cost of also preventing otherwise valid sequential calls.
The state-handling in complete/close also softens the effect of improper/concurrent calls. Thread-safe access and management of state is important for internally-connected 'boundary' sinks/sources, and concurrent access is also intended for sinks/sources shared across a fan-out/in (which can be protected using 'synchronizeStepSource/Sink' if needed). But in other situations, 'close' is not meant to be called concurrently with 'poll'/'drainToSink', and 'complete[Exceptionally]' is not meant to be called concurrently with 'offer'/'drainFromSource'.