Remove flatMap, and rename flatMapStep -> flatMap.
The old flatMap abused a misinterpretation of the drainFromSource contract to identify early cancellation. In particular, a return of 'false' from drainFromSource indicates that the source did not drain, but does not indicate that the sink cancelled (since the sink may internally wrap the source in a new source that drains earlier - as in adaptSourceOfSink). This means that subsequent calls to drainFromSource after a 'false' return may still consume elements and even return 'true'. Therefore, a 'false' return from drainFromSource cannot be used to short-circuit flatMap (unlike a 'false' return from drainToSink, which is what the remaining flatMap uses).