Remove #joinAfterCall, in favor of just letting STS throw a suppressed ISE on close if we throw from its try-block.
This comes after consideration that we would want to capture any exception from STS#join anyway, even if we already threw and needed to suppress the #join exception. But in that case, always forcing a join (in a finally-block) becomes pointless, because we're just trading the suppressed exception that #close would already give us for a different one.