Fix variance for Operators and all chaining methods.
I also removed the pass-through #sink/#stepSink/#source/#stepSource methods, as experientially it seems that encouraging explicit types is going to be a better move than going out of the way to provide helpers to (try to) make inference work. The latter elides useful reading context, exacerbates type-mismatch errors, and encourages direction-swapping (between #andThen and #compose) to maintain inference. For now, it feels like the sanest way to pre-empt most of the compiler-fighting is to just declare the lambda types when there is any doubt.