Remove #spill[Step] for now, as they're unconventional and perhaps niche.
There's also the matter of whether the last element offered (for which a sink returned false) should be re-offered to the next sink or not. Sink#offer cannot convey whether the element was 'fully processed' (whatever that means), 'partially processed', or unprocessed, suggesting that re-offering comes down to choice / an extra parameter. This diverges from its dual, #concat[Step], so keeping #spill[Step] just to complete the duality is especially questionable. For now it seems friendlier to stick to conventional, important, or obviously useful operators.